refactor: ViewHolder의 생성자에서 onClick을 받도록 변경
